PutObjectCommand

+++ src/providers/azureProvider.ts
@@ -9,6 +9,7 @@ import { BlobServiceClient } from "@azure/storage-blob";
 import { DefaultAzureCredential } from "@azure/identity"
 
 import { DistributedProvider } from './provider.js'
+import { StreamingBlobPayloadInputTypes } from "@smithy/types";
 
 async function streamToBuffer(readableStream: NodeJS.ReadableStream): Promise<Buffer> {
     return new Promise((resolve, reject) => {
@@ -47,15 +48,17 @@ export class AzureProvider extends DistributedProvider {
 
     /************************ BUCKET RELATED FUNCTIONS ************************/
     async createBucket(
-        containerName: string
+        containerName: string | undefined
     ) {
+        if (!containerName) throw new Error('Bucket name not found.');
         const containerClient = this.client.getContainerClient(containerName);
         await containerClient.create();
     }
 
     async deleteBucket(
-        bucketName: string
+        bucketName: string | undefined
     ) {
+        if (!bucketName) throw new Error('Bucket name not found.');
         const containerClient = this.client.getContainerClient(bucketName);
         await containerClient.delete();
     }
@@ -71,15 +74,16 @@ export class AzureProvider extends DistributedProvider {
 
     /************************ OBJECT RELATED FUNCTIONS ************************/
     async putObject(
-        bucketName: string,
-        objectKey: string,
-        objectBody: string
+        bucketName: string | undefined,
+        objectKey: string | undefined,
+        objectBody: StreamingBlobPayloadInputTypes | undefined
     ): Promise<void> {
+        if (!bucketName || !objectKey || !objectBody) throw new Error('Bucket name, object key or object body not found.');
         const blockBlobClient = this.client
             .getContainerClient(bucketName)
             .getBlockBlobClient(objectKey);
         // Upload data to the object ( = blob)
-        await blockBlobClient.upload(objectBody, objectBody.length);
+        await blockBlobClient.upload(objectBody, objectBody.toString().length);
     }
 
     // This function might come in handy later but is currently not necessary.
@@ -94,11 +98,10 @@ export class AzureProvider extends DistributedProvider {
     // }
 
     async getObject(
-        bucketName: string,
-        objectKey: string
+        bucketName: string | undefined,
+        objectKey: string | undefined
     ): Promise<string> {
-
-
+        if (!bucketName || !objectKey) throw new Error('Bucket name or object key not found.');
         const blobClient = this.client
             .getContainerClient(bucketName)
             .getBlobClient(objectKey);
@@ -115,9 +118,10 @@ export class AzureProvider extends DistributedProvider {
     }
 
     async deleteObject(
-        bucketName: string,
-        objectKey: string
+        bucketName: string | undefined,
+        objectKey: string | undefined
     ) {
+        if (!bucketName || !objectKey) throw new Error('Bucket name or object key not found.');
         const blockBlobClient = this.client
             .getContainerClient(bucketName)
             .getBlockBlobClient(objectKey);
@@ -125,8 +129,9 @@ export class AzureProvider extends DistributedProvider {
     }
 
     async listObjects(
-        bucketName: string
+        bucketName: string | undefined
     ): Promise<void> {
+        if (!bucketName) throw new Error('Bucket name not found.');
         const blockBlobClient = this.client
             .getContainerClient(bucketName)
         var log = `${bucketName} objects in Azure:\n`;
